+++ packages/codegen.go/src/gomod.ts
@@ -20,7 +20,7 @@ export async function generateGoModFile(codeModel: go.CodeModel, existingGoMod?:
 
   // here we specify the minimum version of azcore as required by the code generator.
   // the version can be overwritten by passing the --azcore-version switch during generation.
-  let version = '1.17.0';
+  let version = '1.17.1';
   if (codeModel.options.azcoreVersion) {
     // when matching versions, we need to handle beta, non-beta, and pseudo versions
     // 1.2.3-beta.1, 1.2.3, 0.22.1-0.20220315231014-ed309e73db6b
@@ -34,7 +34,7 @@ export async function generateGoModFile(codeModel: go.CodeModel, existingGoMod?:
   if (!existingGoMod) {
     // no preexisting go.mod file, generate the default one
     let text = `module ${modName}\n\n`;
-    text += 'go 1.18\n\n';
+    text += 'go 1.23.0\n\n';
     text += `require ${azcore}\n`;
     return text;
   }