+++ packages/typespec-go/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bugs Fixed
 
 * Unsupported tsp constructs and other errors are now reported as a diagnostic error instead of an unhandled exception.
+* Changed the default value of the `--factory-gather-all-params` switch from `false` to `true`.
 
 ### Other Changes
 
@@ -22,6 +23,7 @@
 
 * Added switch `generate-samples` to control example code generation.
 * Deprecated `generate-examples`, use `generate-samples` instead.
+* Changed the default value of the `--factory-gather-all-params` switch from `false` to `true`.
 
 ## 0.3.10 (2025-03-06)